backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0

# Optional (only needed if using WebSockets)
#Flask-SocketIO==5.3.0

openai==1.3.0
firecrawl-py==0.0.16

psycopg2-binary==2.9.9
pgvector==0.2.0

python-dotenv==1.0.0
SQLAlchemy==2.0.23
